Refacting tests
