+++ services/api-server/tests/unit/test_api/conftest.py
@@ -5,28 +5,30 @@
 import pytest
 from starlette.testclient import TestClient
 
-from simcore_service_api_server.__version__ import api_version, api_vtag
-from simcore_service_api_server.main import init_application
+
+from fastapi import FastAPI
 
 
 @pytest.fixture
-def client(monkeypatch) -> TestClient:
+def app(monkeypatch) -> FastAPI:
     monkeypatch.setenv("POSTGRES_USER", "test")
     monkeypatch.setenv("POSTGRES_PASSWORD", "test")
     monkeypatch.setenv("POSTGRES_DB", "test")
     monkeypatch.setenv("LOGLEVEL", "debug")
     monkeypatch.setenv("SC_BOOT_MODE", "production")
 
-    # app
+    # NOTE: keep import inside so monkey-patch has an effect
+    from simcore_service_api_server.main import init_application
     app = init_application()
 
+    return app
+
+
+@pytest.fixture
+def client(app: FastAPI) -> TestClient:
+
+    import pdb; pdb.set_trace()
     # test client:
     # Context manager to trigger events: https://fastapi.tiangolo.com/advanced/testing-events/
     with TestClient(app) as cli:
         yield cli
-
-
-def test_read_service_meta(client: TestClient):
-    response = client.get(f"{api_vtag}/meta")
-    assert response.status_code == 200
-    assert response.json()["version"] == api_version
