+++ services/api-server/tests/unit/test_api/test_user.py
@@ -10,21 +10,7 @@ from starlette import status
 from starlette.testclient import TestClient
 
 from simcore_service_api_server.__version__ import api_version, api_vtag
-from simcore_service_api_server.core import application
-from simcore_service_api_server.core.settings import AppSettings
-from simcore_service_api_server.main import init_application
 
-
-@pytest.fixture(scope="module")
-async def postgres_service(monkeypatch) -> Engine:
-    # test environs
-    monkeypatch.setenv("POSTGRES_USER", "test")
-    monkeypatch.setenv("POSTGRES_PASSWORD", "test")
-    monkeypatch.setenv("POSTGRES_DB", "test")
-
-    # TODO: start pg db and return an engine
-    async with create_engine() as engine:
-        yield engine
 
 
 @pytest.fixture(scope="module")
@@ -40,18 +26,7 @@ async def api_keys_in_db(postgres_service):
     return api_keys
 
 
-@pytest.fixture
-def client(monkeypatch, postgres_service) -> TestClient:
-    # test environs
-    monkeypatch.setenv("LOGLEVEL", "debug")
-    monkeypatch.setenv("SC_BOOT_MODE", "production")
 
-    app = init_application()
-
-    # test client:
-    # Context manager to trigger events: https://fastapi.tiangolo.com/advanced/testing-events/
-    with TestClient(app) as cli:
-        yield cli
 
 
 def test_get_user(client: TestClient, api_keys_in_db: Dict, user_in_db: Dict):
